+++ src/inspect_ai/model/_providers/openai_responses.py
@@ -148,25 +148,29 @@ async def generate_responses(
         return ModelOutput(
             model=model_response.model,
             choices=choices,
-            usage=(
-                ModelUsage(
-                    input_tokens=model_response.usage.input_tokens,
-                    output_tokens=model_response.usage.output_tokens,
-                    input_tokens_cache_read=(
-                        model_response.usage.input_tokens_details.cached_tokens
-                    ),
-                    reasoning_tokens=model_response.usage.output_tokens_details.reasoning_tokens,
-                    total_tokens=model_response.usage.total_tokens,
-                )
-                if model_response.usage
-                else None
-            ),
+            usage=model_usage_from_response(model_response),
         ), model_call()
     except BadRequestError as e:
         if handle_bad_request:
             return handle_bad_request(e), model_call()
         else:
             return openai_handle_bad_request(model_name, e), model_call()
+
+
+def model_usage_from_response(model_response: Response) -> ModelUsage | None:
+    return (
+        ModelUsage(
+            input_tokens=model_response.usage.input_tokens,
+            output_tokens=model_response.usage.output_tokens,
+            input_tokens_cache_read=(
+                model_response.usage.input_tokens_details.cached_tokens
+            ),
+            reasoning_tokens=model_response.usage.output_tokens_details.reasoning_tokens,
+            total_tokens=model_response.usage.total_tokens,
+        )
+        if model_response.usage
+        else None
+    )
 
 
 async def wait_for_background_response(